requirements.txt
appdirs==1.4.3
apturl==0.5.2
arrow==1.2.2
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.7.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
binaryornot==0.4.4
black==22.6.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.0
click==8.0.3
command-not-found==0.3
cookiecutter==2.1.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.2
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.2
fasteners==0.14.1
filelock==3.4.0
flake8==5.0.4
flake8-black==0.3.3
folder-color-common==0.0.86
future==0.18.2
GDAL==3.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
protobuf==3.6.1
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.9.1
pycups==1.9.73
pyflakes==2.5.0
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyngrok==5.1.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-slugify==6.1.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sentry-sdk==1.9.8
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
stack-data==0.1.4
systemd-python==234
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.10.0
wadllib==1.3.3
wcwidth==0.2.5
xkit==0.0.0
zipp==1.0.0
